+++ src/Component.context.ts
@@ -1,5 +1,6 @@
 import { createSafeContext } from '@mantine/core';
+import { ComponentContext } from './Component.types';
 
-export const [ComponentProvider, useComponentContext] = createSafeContext(
+export const [ComponentProvider, useComponentContext] = createSafeContext<ComponentContext>(
   'Component was not found in tree'
 );
